feat: moveit_rviz param
