+++ src/we_r2_arm_moveit/launch/moveit.launch.py
@@ -1,69 +1,31 @@
 import os
 from launch import LaunchDescription
-from moveit_configs_utils import MoveItConfigsBuilder
 from launch_ros.actions import Node
-from launch.actions import DeclareLaunchArgument
-from launch.substitutions import LaunchConfiguration
+from moveit_configs_utils import MoveItConfigsBuilder
 from ament_index_python.packages import get_package_share_directory
 
-
 def generate_launch_description():
-
-    is_sim = LaunchConfiguration("is_sim")
-    
-    is_sim_arg = DeclareLaunchArgument(
-        "is_sim",
-        default_value="True"
-    )
-
     moveit_config = (
         MoveItConfigsBuilder("we_r2_arm", package_name="we_r2_arm_moveit")
         .robot_description(file_path=os.path.join(
             get_package_share_directory("we_r2_arm_description"),
-            "urdf",
-            "we_r2_arm.urdf.xacro"
+                "urdf",
+                "we_r2_arm.urdf.xacro"
             )
         )
         .robot_description_semantic(file_path="config/we_r2_arm.srdf")
         .trajectory_execution(file_path="config/moveit_controllers.yaml")
+        .moveit_cpp(file_path="config/planning_python_api.yaml")
         .to_moveit_configs()
     )
 
-    move_group_node = Node(
-        package="moveit_ros_move_group",
-        executable="move_group",
-        output="screen",
-        parameters=[moveit_config.to_dict(), 
-                    {"use_sim_time": is_sim},
-                    {"publish_robot_description_semantic": True}],
-        arguments=["--ros-args", "--log-level", "info"],
+    move_joint_group_server = Node(
+        package="we_r2_arm_moveit",
+        executable="move_joint_group_server.py",
+        parameters=[moveit_config.to_dict(),
+                    {"use_sim_time": True}]
     )
-
-    # RViz
-    rviz_config = os.path.join(
-        get_package_share_directory("arduinobot_moveit"),
-            "rviz",
-            "moveit.rviz",
-    )
-
-    rviz_node = Node(
-        package="rviz2",
-        executable="rviz2",
-        name="rviz2",
-        output="log",
-        arguments=["-d", rviz_config],
-        parameters=[
-            moveit_config.robot_description,
-            moveit_config.robot_description_semantic,
-            moveit_config.robot_description_kinematics,
-            moveit_config.joint_limits,
-        ],
-    )
-
-    return LaunchDescription(
-        [
-            is_sim_arg,
-            move_group_node, 
-            rviz_node
-        ]
-    )
+    
+    return LaunchDescription([
+        move_joint_group_server,
+    ])
